+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/FirDumpTestGenerated.java
@@ -26,6 +26,18 @@ public class FirDumpTestGenerated extends AbstractFirDumpTest {
   }
 
   @Test
+  @TestMetadata("bindsMirrorInExternallyGeneratedContribution_k23x.kt")
+  public void testBindsMirrorInExternallyGeneratedContribution_k23x() {
+    run("bindsMirrorInExternallyGeneratedContribution_k23x.kt");
+  }
+
+  @Test
+  @TestMetadata("bindsMirrorInExternallyGeneratedContribution_k24x.kt")
+  public void testBindsMirrorInExternallyGeneratedContribution_k24x() {
+    run("bindsMirrorInExternallyGeneratedContribution_k24x.kt");
+  }
+
+  @Test
   @TestMetadata("contributionProvidersShouldNotGenerateForContributesTo.kt")
   public void testContributionProvidersShouldNotGenerateForContributesTo() {
     run("contributionProvidersShouldNotGenerateForContributesTo.kt");